+++ src/main/java/racingcar/dto/CarDto.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.dto;
 
 /**
  * CarDTO.java
